added next auth

+++ app/api/issues/[id]/route.ts
@@ -1,12 +1,16 @@
-import { Issue } from '@prisma/client';
+import { Issue } from "@prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 import { issueSchema } from "@/app/validationSchema";
 import prisma from "@/prisma/client";
-import delay from 'delay';
+import delay from "delay";
+import { getServerSession } from "next-auth";
+import authOptions from "@/app/auth/authOptions";
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
   const body = await request.json();
   const validation = issueSchema.safeParse(body);
   if (!validation.success) return NextResponse.json(validation.error.format());
@@ -27,12 +31,16 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const issue = await prisma.issue.findUnique({ where: { id: parseInt(params.id) } });
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
   if (!issue)
     return NextResponse.json({ error: "invalid issue Id" }, { status: 400 });
 
   await prisma.issue.delete({
     where: { id: issue.id },
   });
-  return NextResponse.json({successfull:"ok"},{ status: 200 });
+  return NextResponse.json({ successfull: "ok" }, { status: 200 });
 }
